requirements.txt
absl-py==0.7.1
args==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
click==6.7
clint==0.5.1
cycler==0.10.0
decorator==4.4.0
future==0.17.1
grpcio==1.21.1
h5py==2.9.0
hestia==0.3.0
idna==2.8
Jinja2==2.10
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.0.0rc4
matplotlib==3.1.0
more-itertools==7.0.0
networkx==2.3
numpy==1.16.3
pandas==0.24.2
pathlib==1.0.1
Pillow==6.0.0
pluggy==0.11.0
polyaxon==0.0.0
polyaxon-cli==0.4.4
polyaxon-client==0.4.4
polyaxon-deploy==0.4.4
polyaxon-schemas==0.4.4
polystores==0.1.7
protobuf==3.8.0
psutil==5.4.7
py==1.8.0
pydot==1.4.1
pyparsing==2.4.0
pytest==4.5.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
raven==6.7.0
requests==2.22.0
requests-toolbelt==0.8.0
rhea==0.4.5
six==1.11.0
tabulate==0.8.2
tb-nightly==1.15.0a20190624
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
urllib3==1.25.2
wcwidth==0.1.7
websocket-client==0.56.0
